backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic-settings==2.5.2
pydantic[email]==2.9.2
pyjwt==2.9.0
bcrypt==4.2.0
boto3==1.35.0
python-multipart==0.0.12
pytest==8.3.3
httpx==0.27.2
pytest-cov==5.0.0
moto[s3]==5.0.17
